CQDBasics.jl

Documentation for CQDBasics.jl.

Contents

CQDBase.jl

CQDDataAnalysisBase.jl

CQDBasics.jl

Installation

In a Julia session, run:

julia> using Pkg

julia> Pkg.add(url="https://github.com/XK-Lin/CQDBasics.jl")

Examples

Examples are here. You may want to create the following directory layout.

Example/
├── CQDDataAnalysis.jl
├── Loop1/
│   ├── CQDLoop.jl
│   └── ...
├── Loop2/
│   ├── CQDLoop.jl
│   └── ...
...

Functions

CQDBasics.CQDBasics — Module

This module wraps CQDBase.jl, CQDDataAnalysisBase.jl, and QuantumMechanics.jl.

Author: Xukun Lin

Update: 10/25/2024

Required package (in addition to those required by wrapped modules): "Reexport".

CQDBasics.get_quantum_mechanics_results — Method

get_quantum_mechanics_results(experiment::Experiment, initial_state::Union{String, Float64}, magnetic_field_computation_method::String, sigmoid_field::Union{String, Tuple{<:Real, <:Real}}, show_plot::Bool)

Calculate the quantum mechanical predictions of the spin flip probabilities.

CQDBasics.see_magnetic_fields — Method

see_magnetic_fields(plot_range::Tuple{<:Real, <:Real}, current::Real, zₐ::Real, Bᵣ::Vector{<:Real}, magnetic_field_computation_method::String, sigmoid_field::Union{String, Tuple{<:Real, <:Real}})

Visualize the z component of the magnetic fields.

Index

CQDBasics.CQDBase

CQDBasics.CQDBasics

CQDBasics.CQDDataAnalysisBase

CQDBasics.CQDBase.Experiment

CQDBasics.CQDBase.Experiment

CQDBasics.CQDBase.Results

CQDBasics.CQDBase.Results

CQDBasics.CQDBase.Simulation

CQDBasics.CQDBase.CQD_LLG_equation!

CQDBasics.CQDBase.apply_branching

CQDBasics.CQDBase.apply_branching

CQDBasics.CQDBase.cartesian_to_spherical

CQDBasics.CQDBase.get_external_magnetic_fields

CQDBasics.CQDBase.get_external_magnetic_fields

CQDBasics.CQDBase.get_external_magnetic_fields_at_ends

CQDBasics.CQDBase.get_perpendicular_norm_vector

CQDBasics.CQDBase.get_solver

CQDBasics.CQDBase.is_flipped

CQDBasics.CQDBase.is_flipped

CQDBasics.CQDBase.latex_exponential

CQDBasics.CQDBase.sample_atom_once

CQDBasics.CQDBase.sample_atom_once

CQDBasics.CQDBase.sample_atoms

CQDBasics.CQDBase.sample_atoms

CQDBasics.CQDBase.save_results

CQDBasics.CQDBase.simulate

CQDBasics.CQDBase.spherical_to_cartesian

CQDBasics.CQDBase.transform_angles

CQDBasics.CQDBase.transform_vectors

CQDBasics.CQDBase.wrap

CQDBasics.CQDDataAnalysisBase.categorize_folders

CQDBasics.CQDDataAnalysisBase.clean_folders

CQDBasics.CQDDataAnalysisBase.combine_excels

CQDBasics.CQDDataAnalysisBase.get_kᵢ_plots

CQDBasics.CQDDataAnalysisBase.get_subfolder_summary_tables

CQDBasics.CQDDataAnalysisBase.sort_folders

CQDBasics.CQDDataAnalysisBase.unpack_folders

CQDBasics.get_quantum_mechanics_results

CQDBasics.see_magnetic_fields