CQDBase

CQDBasics.CQDBase.Experiment — Type

struct Experiment

An Experiment represents the experiment to simulate.

Fields

name::String: The name of the experiment. Predefined experiment names are "FS Low $z_a$", and "FS High $z_a$", "Alex 165", "Alex 156", "Alex 396", "Alex 105". For predefined experiments, use Experiment(name).

currents::Vector{<:Real}: The wire currents.

flip_probabilities::Vector{<:Real}: The (mean) flip probabilities from the experiment.

flip_probabilities_stds::Vector{<:Real}: The standard deviations of the flip probabilities.

qm_flip_probabilities::Vector{<:Real}: The flip probabilities according to QM simulation.

zₐ::Real: The distance from the beam to the null point.

v::Real: The velocity of the atomic beam.

Bᵣ::Vector{<:Real}: The remnant field.

system_length::Real: The length of the system.

time_span::Tuple{<:Real, <:Real}: The flight time range for the atoms.

CQDBasics.CQDBase.Experiment — Method

Experiment(name::String)

Outer constructor for Experiment that takes a predefined experiment name.

CQDBasics.CQDBase.Results — Type

struct Results

A Results represents the results of the simulation.

Fields

raw_data::BitArray: The raw simulation data of whether the atoms flip.

flip_probabilities::Vector{<:Real}: The flip probability calculated from the simulation.

flip_probabilities_stds::Vector{<:Real}: The standard deviations of the calculated flip probabilities.

R2::Real: The R sqaure value calculated from the simulation and experiment.

θₑ_plot::Plots.Plot: The plot of θₑ dynamics.

θₙ_plot::Plots.Plot: The plot of θₙ dynamics.

θₑθₙ_plot::Plots.Plot: The plot that combines the two θ plots.

flip_plot::Plots.Plot: The plot of the simulation and experiment flip probabilities.

CQDBasics.CQDBase.Results — Method

Results(experiment::Experiment, simulation::Simulation, raw_data::BitArray, θₑ_plot::Plots.Plot, θₙ_plot::Plots.Plot, θₑθₙ_plot::Plots.Plot)

Outer constructor for Results.

CQDBasics.CQDBase.Simulation — Type

struct Simulation

A Simulation represents a particular simulation setup.

Fields

type::String: The type of simulation. Choose from BE and WM.

atom_number::Int: The number of atoms.

magnetic_field_computation_method::String: The method used for calculating the magnetic field due to the wire. Choose from "quadrupole" and "exact".

initial_μₑ::Union{String, Real}: The initial condition for the electron magnetic moments. Choose from "up", "down", 1/2, or -1/2.

initial_μₙ::Union{String, Vector{<:Real}}: The sampling method for nuleus magnetic moments. Choose from "HS", "HS 2", "HS 4", "IHS", "IHS 2", "IHS 4", "Iso", "Iso 2", "Iso 4", or give a vector of probability weights. The vector should be of length 2 or 4.

solver::String: The differential equation solver. Several good ones are "radau", "radau5", "RadauIIA5", and "TRBDF2".

θₙ_is_fixed::Bool: Whether θₙ is fixed or not.

branching_condition::String: Which branching condition to use. Choose from "B₀ dominant" and "Bₑ dominant".

BₙBₑ_strength::String: The values for Bₙ and Bₑ. Choose from "CQD" and "quantum".

BₙBₑ_ratio::Tuple{<:Real, <:Real}: The ratio of the used Bₙ and Bₑ to the theory value.

kᵢ::Real: The collapse coefficient.

average_method::Union{String, Tuple{String, <:Real}}: The average method. Choose from "ABC" (average angles then branching condition), "BCA" (branching conditions averaged), and "no average". If using "ABC" or "BCA", the input should be a tuple of length 2, where the second entry is the fraction of total time to be averaged. For example, average_method = ("ABC", 1/16).

θ_cross_detection::Union{String, Tuple{String, <:Real, <:Real}, Tuple{String, <:Real, String}} (currently disabled): Whether angle cross is automatically detected. Choose from "off", "sign", and "minabs". If using "sign" or "minabs", the input should be a tuple of length 3, where the second entry is the start time for detection, and the third entry is the period for detection. The period may be a fixed value or "adaptive". For example, θ_cross_detection = ("sign", 10e-6, 4.5e-6).

sigmoid_field::Union{String, Tuple{<:Real, <:Real}}: Whether use a sigmoid transition field. Give either "off" or a tuple of length 2, where the first entry is the magnetic field strength, and the second entry is the y coordinate of the SG apparatus. For example, sigmoid_field = (0.1, 2e-2).

R2_comparison::String: How to calculate R2. Choose from "experiment" and "qm".

save_files::Union{String, Vector{String}}: Which files to save. Choose from "θₑ plot", "θₙ plot", "θₑ θₙ plot", "flip plot", "raw data", "flip probabilities", "CQDBase.jl", "simulation info", and "package info".

CQDBasics.CQDBase.CQD_LLG_equation! — Method

CQD_LLG_equation!(du, u, p, t)

Define the differential equation using CQD LLG equations.

Arguments

u = [θₑ, θₙ, ϕₑ, ϕₙ]: The variable of the differential equation. p = [experiment::Experiment, simulation::Simulation, current, Bₙ, Bₑ]: The parameters passed to the solver.

CQDBasics.CQDBase.apply_branching — Method

apply_branching(θₑs::Union{Real, Vector{<:Real}}, θₙs::Union{Real, Vector{<:Real}}, ϕₑs::Union{Real, Vector{<:Real}}, ϕₙs::Union{Real, Vector{<:Real}}, mᵢ::Real, simulation::Simulation)

Apply the branching conditions.

CQDBasics.CQDBase.cartesian_to_spherical — Method

cartesian_to_spherical(x::Real, y::Real, z::Real)

Convert Cartesian coordinates (x, y, z) to spherical coordinates (r, θ, ϕ).

CQDBasics.CQDBase.get_external_magnetic_fields — Method

get_external_magnetic_fields(t::Real, current::Real, experiment::Experiment, simulation::Simulation)

Calculate the external magnetic field components Bx, By, Bz at a given time t and return a tuple.

Notes

+y is right, +z is up, +x is out of page.

Due to a different definition of the sign of current, the expressions have an extra minus sign.

CQDBasics.CQDBase.get_external_magnetic_fields — Method

get_external_magnetic_fields(t::Real, current::Real, experiment::Experiment, magnetic_field_computation_method::String, sigmoid_field::Union{String, Tuple{<:Real, <:Real}})

Calculate the external magnetic field components Bx, By, Bz at a given time t and return a tuple.

Notes

+y is right, +z is up, +x is out of page.

Due to a different definition of the sign of current, the expressions have an extra minus sign.

CQDBasics.CQDBase.get_external_magnetic_fields_at_ends — Method

get_external_magnetic_fields_at_ends(current::Real, experiment::Experiment, simulation::Simulation)

Calculate the external magnetic field components at the two ends of the flipper.

CQDBasics.CQDBase.get_perpendicular_norm_vector — Method

get_perpendicular_norm_vector(v::Vector{<:Real})

Find one normalized vector that is perpendicular to v.

CQDBasics.CQDBase.get_solver — Method

get_solver(name::String)

Get the solver corresponding to the name.

CQDBasics.CQDBase.is_flipped — Method

is_flipped(angles::Union{Vector{<:Real}, Vector{<:Vector{<:Real}}}, mᵢ::Real, current::Real, experiment::Experiment, simulation::Simulation)

Determine whether an atom has flipped based on its angles, magnetic field computation method, and the branching condition.

CQDBasics.CQDBase.latex_exponential — Method

latex_exponential(x::Real)

Convert a number x to a beautiful scientific-notation latex string.

CQDBasics.CQDBase.sample_atom_once — Method

sample_atom_once(simulation::Simulation)

Sample one atom based on the simulation parameters.

CQDBasics.CQDBase.sample_atoms — Method

sample_atoms(simulation::Simulation)

Sample atoms according to the predefined number of atoms.

CQDBasics.CQDBase.save_results — Method

save_results(experiment::Experiment, simulation::Simulation, results::Results, start_time, file_dir)

Save the results of the whole simulation.

CQDBasics.CQDBase.simulate — Method

simulate(experiment::Experiment, simulation::Simulation)

Simulate the whole system.

CQDBasics.CQDBase.spherical_to_cartesian — Method

spherical_to_cartesian(r::Real, θ::Real, ϕ::Real)

Convert spherical coordinates (r, θ, ϕ) to Cartesian coordinates (x, y, z).

CQDBasics.CQDBase.transform_angles — Method

transform_angles(θ::Union{Real, Vector{<:Real}}, ϕ::Union{Real, Vector{<:Real}}, z₁::Vector{<:Real}, z₂::Vector{<:Real})

Express θ and ϕ, which are defined in the z₁ spherical coordinates, in the z₂ spherical coordinates.

Notes

There are infinitly many possible rotations. Here only the simpliest case is considered.

CQDBasics.CQDBase.transform_vectors — Method

transform_vectors(vectors::Union{Vector{Vector{<:Real}}, Vector{<:Real}})

Transform a vector (length M) of vectors (length N) to a vector of length N with elements of length M.

If the input is a vector but not a vector of vectors, return the vector itself.

CQDBasics.CQDBase.wrap — Method

wrap(θ)

Wrap the angle θ to between 0 and π.