CQDDataAnalysisBase

CQDBasics.CQDBase.Experiment

CQDBasics.CQDBase.Experiment

CQDBasics.CQDBase.Results

CQDBasics.CQDBase.Results

CQDBasics.CQDBase.Simulation

CQDBasics.CQDBase.CQD_LLG_equation!

CQDBasics.CQDBase.apply_branching

CQDBasics.CQDBase.cartesian_to_spherical

CQDBasics.CQDBase.get_external_magnetic_fields

CQDBasics.CQDBase.get_external_magnetic_fields

CQDBasics.CQDBase.get_external_magnetic_fields_at_ends

CQDBasics.CQDBase.get_perpendicular_norm_vector

CQDBasics.CQDBase.get_solver

CQDBasics.CQDBase.is_flipped

CQDBasics.CQDBase.latex_exponential

CQDBasics.CQDBase.sample_atom_once

CQDBasics.CQDBase.sample_atoms

CQDBasics.CQDBase.save_results

CQDBasics.CQDBase.simulate

CQDBasics.CQDBase.spherical_to_cartesian

CQDBasics.CQDBase.transform_angles

CQDBasics.CQDBase.transform_vectors

CQDBasics.CQDBase.wrap

CQDBasics.CQDDataAnalysisBase.categorize_folders

CQDBasics.CQDDataAnalysisBase.clean_folders

CQDBasics.CQDDataAnalysisBase.combine_excels

CQDBasics.CQDDataAnalysisBase.get_kᵢ_plots

CQDBasics.CQDDataAnalysisBase.get_subfolder_summary_tables

CQDBasics.CQDDataAnalysisBase.sort_folders

CQDBasics.CQDDataAnalysisBase.unpack_folders

CQDBasics.CQDDataAnalysisBase.categorize_folders — Method

categorize_folders(directory::String, fields::Union{String, Tuple{Vararg{String}}})

Divide all folders in directory into groups according to the fields. Then for each group, create a new folder in directory and move the group into it.

CQDBasics.CQDDataAnalysisBase.clean_folders — Method

clean_folders(directory::String, save_original::Bool)

Clean the contents of folders in directory for faster upload / download.

Note that if you are using other functions from CQDDataAnalysisBase.jl in the folder dir, then the input for this function should be dirname(dir).

CQDBasics.CQDDataAnalysisBase.combine_excels — Method

combine_excels(directory::String, sheet_name::String, column_for_average::Int64)

Combine all excel .xlsx files in directory. All excel files are supposed to have sheets with name sheet_name, and the column heads are supposed to be identical. For rows that have identical values for all columns but column #column_for_average, their values in column #column_for_average are averaged.

CQDBasics.CQDDataAnalysisBase.get_kᵢ_plots — Method

get_kᵢ_plots(directory::String)

Plot the data from the folders within the directory.

CQDBasics.CQDDataAnalysisBase.get_subfolder_summary_tables — Method

get_subfolder_summary_tables(directory::String)

For each folder in directory, create an Excel file that contains Max-R2-ranked configurations.

CQDBasics.CQDDataAnalysisBase.sort_folders — Method

sort_folders(directory::String)

Sort folders in directory based on the max R2 value.

CQDBasics.CQDDataAnalysisBase.unpack_folders — Method

unpack_folders(directory::String)

For each folder in directory, move all subfolders out of it and delete the folder. Files in the folder that are not directories are deleted.